chore: convert validator to yup
